Bugfix/scheduling

+++ src/main/java/com/w2w/api/scheduling/dto/PositionShiftBucket.java
@@ -1,11 +1,12 @@
 package com.w2w.api.scheduling.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record PositionShiftBucket(
         String position,
         List<EmployeeShift> shifts,
         Integer shiftCount,
-        Float totalDuration
+        BigDecimal totalDuration
 ) {
 }
